Email

+++ src/main/java/com/tabber/tabby/service/impl/ContestWidgetServiceImpl.java
@@ -3,16 +3,12 @@ package com.tabber.tabby.service.impl;
 import com.tabber.tabby.dto.ContestWidgetRequest;
 import com.tabber.tabby.entity.ContestWidgetEntity;
 import com.tabber.tabby.entity.UserEntity;
-import com.tabber.tabby.exceptions.ContestWidgetExistsException;
 import com.tabber.tabby.exceptions.ContestWidgetNotExistsException;
-import com.tabber.tabby.exceptions.RankWidgetNotExistsException;
 import com.tabber.tabby.respository.ContestWidgetRepository;
 import com.tabber.tabby.service.ContestWidgetService;
 import com.tabber.tabby.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 public class ContestWidgetServiceImpl implements ContestWidgetService {
@@ -24,11 +20,8 @@ public class ContestWidgetServiceImpl implements ContestWidgetService {
     ContestWidgetRepository contestWidgetRepository;
 
     @Override
-    public ContestWidgetEntity createContestWidget(ContestWidgetRequest contestWidgetRequest, Long userId) throws ContestWidgetExistsException{
+    public ContestWidgetEntity createContestWidget(ContestWidgetRequest contestWidgetRequest, Long userId) {
         UserEntity userEntity = userService.getUserFromUserId(userId);
-        if(contestWidgetExistsForWebsite(userEntity,contestWidgetRequest)!=null){
-            throw new ContestWidgetExistsException("Exists for user "+userId+" for website id "+contestWidgetRequest.getWebsiteId());
-        }
         ContestWidgetEntity contestWidgetEntity = new ContestWidgetEntity()
                 .toBuilder()
                 .websiteId(contestWidgetRequest.getWebsiteId())
@@ -44,8 +37,11 @@ public class ContestWidgetServiceImpl implements ContestWidgetService {
     }
     @Override
     public ContestWidgetEntity updateContestWidget(ContestWidgetRequest contestWidgetRequest, Long userId) throws ContestWidgetNotExistsException{
-        UserEntity userEntity = userService.getUserFromUserId(userId);
-        ContestWidgetEntity contestWidget = contestWidgetExistsForWebsite(userEntity,contestWidgetRequest);
+        if(contestWidgetRequest.getId() == null){
+            throw new ContestWidgetNotExistsException("Widget Id not specified");
+
+        }
+        ContestWidgetEntity contestWidget = contestWidgetRepository.getTopByWidgetId(contestWidgetRequest.getId());
         if(contestWidget==null){
             throw new ContestWidgetNotExistsException("Doesn't exist for user "+userId+" for website id "+contestWidgetRequest.getWebsiteId());
         }
@@ -60,11 +56,14 @@ public class ContestWidgetServiceImpl implements ContestWidgetService {
     }
 
     @Override
-    public ContestWidgetEntity deleteContestWidget(ContestWidgetRequest contestWidgetRequest, Long userId) throws ContestWidgetNotExistsException {
+    public ContestWidgetEntity deleteContestWidget(Long contestId, Long userId) throws ContestWidgetNotExistsException {
         UserEntity userEntity = userService.getUserFromUserId(userId);
-        ContestWidgetEntity contestWidget = contestWidgetExistsForWebsite(userEntity,contestWidgetRequest);
+        if(contestId == null){
+            throw new ContestWidgetNotExistsException("Widget Id not specified");
+        }
+        ContestWidgetEntity contestWidget = contestWidgetRepository.getTopByWidgetId(contestId);
         if(contestWidget==null){
-            throw new RankWidgetNotExistsException("Doesn't exist for user "+userId+" for website id "+contestWidgetRequest.getWebsiteId());
+            throw new ContestWidgetNotExistsException("Doesn't exist for user "+userId+" for website id "+contestId);
         }
         contestWidgetRepository.delete(contestWidget);
         userEntity.getContestWidgets().remove(contestWidget);
@@ -72,13 +71,5 @@ public class ContestWidgetServiceImpl implements ContestWidgetService {
         return contestWidget;
     }
 
-    private ContestWidgetEntity contestWidgetExistsForWebsite(UserEntity userEntity, ContestWidgetRequest contestWidgetRequest){
-        List<ContestWidgetEntity> contestWidgetEntities = userEntity.getContestWidgets();
-        for(ContestWidgetEntity contestWidget:contestWidgetEntities){
-            if(contestWidget.getWebsiteId().equals(contestWidgetRequest.getWebsiteId())){
-                return contestWidget;
-            }
-        }
-        return null;
-    }
+
 }
